Major Update 0.9.7
Changed key names in companies dictionary so that they are now tickers from stocklist.

Fixed error in scoring for i03 in sniffer.py

Added new methods to Company class that allow individual metrics and metric names to be viewed in the Lab.

Added new functionality and documentation to TheSmelloscope lab.